Unify common interface for sync and async
