+++ core/src/main/scalajvm/sttp/client4/quick.scala
@@ -1,7 +1,7 @@
 package sttp.client4
 
 object quick extends SttpApi {
-  lazy val backend: SyncBackend = DefaultSyncBackend()
+  lazy val backend: WebSocketBackend[Identity] = DefaultSyncBackend()
 
   implicit class RichRequest[T](val request: Request[T]) {
     def send(): Response[T] = backend.send(request)
