+++ core/src/test/scalajvm/sttp/client4/HttpClientSyncHttpTest.scala
@@ -1,15 +1,15 @@
 package sttp.client4
 
 import sttp.client4.httpclient.HttpClientSyncBackend
+import sttp.client4.monad.IdMonad
+import sttp.client4.testing.websocket.WebSocketTest
 import sttp.client4.testing.{ConvertToFuture, HttpTest}
+import sttp.monad.MonadError
 
-class HttpClientSyncHttpTest extends HttpTest[Identity] {
-  override val backend: SyncBackend = HttpClientSyncBackend()
+class HttpClientSyncHttpTest extends WebSocketTest[Identity] {
+  override val backend: WebSocketBackend[Identity] = HttpClientSyncBackend()
   override implicit val convertToFuture: ConvertToFuture[Identity] = ConvertToFuture.id
+  override implicit val monad: MonadError[Identity] = IdMonad
 
-  override def supportsHostHeaderOverride = false
-  override def supportsCancellation: Boolean = false
-  override def supportsDeflateWrapperChecking = false
-
-  override def timeoutToNone[T](t: Identity[T], timeoutMillis: Int): Identity[Option[T]] = Some(t)
+  override def throwsWhenNotAWebSocket: Boolean = true
 }
