+++ core/src/test/scalajvm/sttp/client4/testing/HttpClientSyncWebsocketTest.scala
@@ -1,0 +1,15 @@
+package sttp.client4
+
+import sttp.client4.httpclient.HttpClientSyncBackend
+import sttp.client4.testing.{ConvertToFuture, HttpTest}
+
+class HttpClientSyncWebsocketTest extends HttpTest[Identity] {
+  override val backend: WebSocketBackend[Identity] = HttpClientSyncBackend()
+  override implicit val convertToFuture: ConvertToFuture[Identity] = ConvertToFuture.id
+
+  override def supportsHostHeaderOverride = false
+  override def supportsCancellation: Boolean = false
+  override def supportsDeflateWrapperChecking = false
+
+  override def timeoutToNone[T](t: Identity[T], timeoutMillis: Int): Identity[Option[T]] = Some(t)
+}
