+++ examples/src/main/scala/sttp/client4/examples/GetRawResponseBodySynchronous.scala
@@ -13,7 +13,7 @@ object GetRawResponseBodySynchronous extends App {
     .get(uri"https://httpbin.org/get")
     .response(asBoth(asJson[HttpBinResponse], asStringAlways))
 
-  val backend: SyncBackend = HttpClientSyncBackend()
+  val backend: WebSocketBackend[Identity] = HttpClientSyncBackend()
 
   try {
     val response: Response[(Either[ResponseException[String, circe.Error], HttpBinResponse], String)] =
